loadRequest should check for empty URLs. https://bugs.webkit.org/show_bug.cgi?id=88154
In [WebFrame loadRequest], if the url is invalid and not nil, we transform it to file: url.
However, WebKit client could send down an empty URL that makes us transform it to file: url
as well.  We should skip that as we did for nil URLs.

Patch by Yongjun Zhang <yongjun_zhang@apple.com> on 2012-06-02
Reviewed by Brady Eidson.

* WebView/WebFrame.mm:
(-[WebFrame loadRequest:]):

git-svn-id: svn://svn.chromium.org/blink/trunk@119339 bbb929c8-8fbe-4397-9dbb-9b2b20218538