Canvas 2D - Source rectangles that overlap the source image boundary, not supported by drawImage https://bugs.webkit.org/show_bug.cgi?id=66574
Patch by Justin Novosad <junov@chromium.org> on 2012-06-04
Reviewed by Stephen White.

Source/WebCore:

Test: fast/canvas/drawImage-clipped-source.html

When the source rectangle argument to drawImage is not entirely
contatained by the domain of the source image, the source rectangle
is clipped, and the destination rectangle is clipped proportionately
Old behavior was to silently fail, which did not comply with the spec.

* html/canvas/CanvasRenderingContext2D.cpp:
(WebCore::normalizeAndClipRects):
(WebCore):
(WebCore::CanvasRenderingContext2D::drawImage):

LayoutTests:

Adding a new layout test that verifies that the source and destination
rectangle arguments to drawImage are properly clipped when the source
rectangle is not entirely contained whithin the domian of the source
image. Tests image to canvas draws and canvas to canvas draws.

* fast/canvas/drawImage-clipped-source-expected.txt: Added.
* fast/canvas/drawImage-clipped-source.html: Added.
* fast/canvas/drawImage-clipped-source.js: Added.
(patternTest.this.testPixel):
(patternTest.this.testRedSquare):
(patternTest.this.testAggregatePattern):
(patternTest):
(drawTestPattern):
(executeTest):

git-svn-id: svn://svn.chromium.org/blink/trunk@119393 bbb929c8-8fbe-4397-9dbb-9b2b20218538