[CSSRegions]NamedFlow::overset should return true when there's no region chain attached https://bugs.webkit.org/show_bug.cgi?id=88515
The CSS Regions specifies that the "overset" attribute needs to return true if the content overflows, there's no region attached to the flow
or if the NamedFlow is in the NULL state:
http://www.w3.org/TR/css3-regions/#dom-named-flow-overset

Source/WebCore:

This patch updates WebKit's behavior in the last two situations to match the one in the spec.

Patch by Andrei Bucur <abucur@adobe.com> on 2012-06-12
Reviewed by Andreas Kling.

Test: fast/regions/webkit-named-flow-overset.html was updated

* rendering/RenderFlowThread.cpp:
(WebCore::RenderFlowThread::RenderFlowThread):
(WebCore::RenderFlowThread::computeOverflowStateForRegions): If the last valid region is null, the flow has not region chain attached or it's in the NULL state.

LayoutTests:

The overset test was updated to cover the attribute value in case of a flow in the NULL state.

Patch by Andrei Bucur <abucur@adobe.com> on 2012-06-12
Reviewed by Andreas Kling.

* fast/regions/webkit-named-flow-overset-expected.txt:
* fast/regions/webkit-named-flow-overset.html:

git-svn-id: svn://svn.chromium.org/blink/trunk@120071 bbb929c8-8fbe-4397-9dbb-9b2b20218538