[CSS3 Backgrounds and Borders] Implement box-decoration-break rendering. https://bugs.webkit.org/show_bug.cgi?id=88228
Reviewed by Eric Seidel.

Source/WebCore:

Implement the new CSS property box-decoration-break. It modifies
where we decide whether the borders needs to be sliced or not by
checking if the box-decoration-break is set to clone. If it's the case
then we need to explicitely redraw all edges.

Test: fast/box-decoration-break/box-decoration-break-rendering.html

* rendering/InlineFlowBox.cpp:
(WebCore::InlineFlowBox::determineSpacingForFlowBoxes):
(WebCore::InlineFlowBox::paintFillLayer):

LayoutTests:

Add new tests to cover the feature.

* fast/box-decoration-break/box-decoration-break-rendering-expected.html: Added.
* fast/box-decoration-break/box-decoration-break-rendering.html: Added.


git-svn-id: svn://svn.chromium.org/blink/trunk@120835 bbb929c8-8fbe-4397-9dbb-9b2b20218538