[chromium] CSS tweaks to media controls to prevent styles from being overridden https://bugs.webkit.org/show_bug.cgi?id=91802
Reviewed by Eric Carlson.

Source/WebCore:

Having the buttons declared as inline caused them to not be included in the box layout,
resulting in styles set outside the media element to affect the position of the buttons.

In a similar vein, it was also possible to override the font-style attribute for the time
display.

Test: media/controls-styling-strict.html

* css/mediaControlsChromium.css:
(audio::-webkit-media-controls-mute-button, video::-webkit-media-controls-mute-button):
(audio::-webkit-media-controls-play-button, video::-webkit-media-controls-play-button):
(audio::-webkit-media-controls-time-remaining-display, video::-webkit-media-controls-time-remaining-display):
(audio::-webkit-media-controls-fullscreen-button, video::-webkit-media-controls-fullscreen-button):

LayoutTests:

* media/controls-styling-strict.html: Added.
* platform/chromium-linux/media/controls-styling-strict-expected.png: Added.
* platform/chromium-linux/media/controls-styling-strict-expected.txt: Added.

git-svn-id: svn://svn.chromium.org/blink/trunk@123259 bbb929c8-8fbe-4397-9dbb-9b2b20218538